move all functional tests into one file
