+++ tox.ini
@@ -46,15 +46,15 @@ commands=
 [testenv:snakemake]
 deps=snakemake
 commands =
-    python -m pytest tests/functional/test_snakemake.py
+    python -m pytest tests/functional/test_functional.py::test_snakemake
 
 [testenv:cromwell]
 # Empty deps. Otherwise deps from default testenv are used.
 deps=
 commands =
-    python -m pytest tests/functional/test_wdl.py::test_cromwell
+    python -m pytest tests/functional/test_functional.py::test_cromwell
 
 [testenv:miniwdl]
 deps=miniwdl
 commands=
-    python -m pytest tests/functional/test_wdl.py::test_miniwdl
+    python -m pytest tests/functional/test_functional.py::test_miniwdl
